requirements.txt
google-generativeai
psycopg2-binary
python-docx
pypdf
python-dotenv
nltk